update : sleepとwaitに負の値を入れたときに警告を表示し０に補正するように変更

+++ modules/API/ClockUtil.h
@@ -12,9 +12,9 @@ class ClockUtil {
  public:
   /**
    * @brief 走行時間を取得
-   * @return 走行時間(ミリ秒)
+   * @return 走行時間(秒)
    */
-  static int now();
+  static double now();
 
   /**
    * @brief 自タスクスリープ（デフォルトは10ミリ秒）
